# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Plant Control Monitoring API",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/users/{role}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "get list user by role",
                "operationId": "fetch user",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "description": "you need login first before using this api",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful"
                    },
                    "401": {
                        "description": "unauthenticated"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "get user detail by id",
                "operationId": "fetch user by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "you need login first before using this api",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful"
                    },
                    "401": {
                        "description": "unauthenticated"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "edit profile user (note : bisa isi salah satu, dua,dst atau semua attribut)",
                "operationId": "editProfilUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "type": "integer"
                    },
                    {
                        "name": "foto",
                        "in": "formData",
                        "description": "foto user",
                        "type": "file"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "nama profil",
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "email user",
                        "type": "string"
                    },
                    {
                        "name": "no_hp",
                        "in": "formData",
                        "description": "nomor hp user",
                        "type": "string"
                    },
                    {
                        "name": "alamat",
                        "in": "formData",
                        "description": "alamat",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}